use ssize_t instead of int

+++ cupy/core/carray.cuh
@@ -135,7 +135,7 @@ __device__ float16 nextafter(float16 x, float16 y) {return float16::nextafter(x,
 
 // CArray
 #define CUPY_FOR(i, n) \
-    for (int i = blockIdx.x * blockDim.x + threadIdx.x; \
+    for (ssize_t i = blockIdx.x * blockDim.x + threadIdx.x; \
          i < (n); \
          i += blockDim.x * gridDim.x)
 
@@ -143,24 +143,24 @@ template <typename T, int ndim>
 class CArray {
 private:
   T* data_;
-  int size_;
-  int shape_[ndim];
-  int strides_[ndim];
+  ssize_t size_;
+  ssize_t shape_[ndim];
+  ssize_t strides_[ndim];
 
 public:
-  __device__ int size() const {
+  __device__ ssize_t size() const {
     return size_;
   }
 
-  __device__ const int* shape() const {
+  __device__ const ssize_t* shape() const {
     return shape_;
   }
 
-  __device__ const int* strides() const {
+  __device__ const ssize_t* strides() const {
     return strides_;
   }
 
-  __device__ T& operator[](const int* idx) {
+  __device__ T& operator[](const ssize_t* idx) {
     char* ptr = reinterpret_cast<char*>(data_);
     for (int dim = 0; dim < ndim; ++dim) {
       ptr += strides_[dim] * idx[dim];
@@ -168,11 +168,11 @@ public:
     return *reinterpret_cast<T*>(ptr);
   }
 
-  __device__ T operator[](const int* idx) const {
+  __device__ T operator[](const ssize_t* idx) const {
     return (*const_cast<CArray<T, ndim>*>(this))[idx];
   }
 
-  __device__ T& operator[](int i) {
+  __device__ T& operator[](ssize_t i) {
     char* ptr = reinterpret_cast<char*>(data_);
     for (int dim = ndim; --dim > 0; ) {
       ptr += strides_[dim] * (i % shape_[dim]);
@@ -185,7 +185,7 @@ public:
     return *reinterpret_cast<T*>(ptr);
   }
 
-  __device__ T operator[](int i) const {
+  __device__ T operator[](ssize_t i) const {
     return (*const_cast<CArray<T, ndim>*>(this))[i];
   }
 };
@@ -194,26 +194,26 @@ template <typename T>
 class CArray<T, 0> {
 private:
   T* data_;
-  int size_;
+  ssize_t size_;
 
 public:
-  __device__ int size() const {
+  __device__ ssize_t size() const {
     return size_;
   }
 
-  __device__ T& operator[](const int* idx) {
+  __device__ T& operator[](const ssize_t* idx) {
     return *reinterpret_cast<T*>(data_);
   }
 
-  __device__ T operator[](const int* idx) const {
+  __device__ T operator[](const ssize_t* idx) const {
     return (*const_cast<CArray<T, 0>*>(this))[idx];
   }
 
-  __device__ T& operator[](int i) {
+  __device__ T& operator[](ssize_t i) {
     return *reinterpret_cast<T*>(data_);
   }
 
-  __device__ T operator[](int i) const {
+  __device__ T operator[](ssize_t i) const {
     return (*const_cast<CArray<T, 0>*>(this))[i];
   }
 };
@@ -221,19 +221,19 @@ public:
 template <int ndim>
 class CIndexer {
 private:
-  int size_;
-  int shape_[ndim];
-  int index_[ndim];
+  ssize_t size_;
+  ssize_t shape_[ndim];
+  ssize_t index_[ndim];
 
 public:
-  __device__ int size() const {
+  __device__ ssize_t size() const {
     return size_;
   }
 
-  __device__ void set(int i) {
-    unsigned int a = i;
-    for (int dim = ndim; --dim > 0; ) {
-      unsigned int s = shape_[dim];
+  __device__ void set(ssize_t i) {
+    ssize_t a = i;
+    for (ssize_t dim = ndim; --dim > 0; ) {
+      ssize_t s = shape_[dim];
       index_[dim] = (a % s);
       a /= s;
     }
@@ -242,7 +242,7 @@ public:
     }
   }
 
-  __device__ const int* get() const {
+  __device__ const ssize_t* get() const {
     return index_;
   }
 };
@@ -250,17 +250,17 @@ public:
 template <>
 class CIndexer<0> {
 private:
-  int size_;
+  ssize_t size_;
 
 public:
-  __device__ int size() const {
+  __device__ ssize_t size() const {
     return size_;
   }
 
-  __device__ void set(int i) {
+  __device__ void set(ssize_t i) {
   }
 
-  __device__ const int* get() const {
+  __device__ const ssize_t* get() const {
     return NULL;
   }
 };
